Extracted generic validation pipeline behaviour out of validation of CreateContactCommand
This allows us to have any command for which we have authored a `FluentValidation` validator be validated. Previously the pipeline behaviour was very tied to the specific validator. For now we have assumed validation will be asynchronous because there does not appear to be an obvious way to tell whether a validator has any async rules.

We _could_ accomplish this by plugging `FluentValidation` directly into ASP.NET Core. However

* That does not support asynchronous validation
* Our request might be pulling in data from multiple places
* We might not even be in ASP.NET Core in other scenarios
* This is a learning exercise!
